Truncate newdemo size to unsigned long for Windows

+++ main/newdemo.cpp
@@ -3299,7 +3299,7 @@ void newdemo_playback_one_frame()
 					num_objs = Highest_object_index;
 					cur_objs = (dxxobject *)d_malloc(sizeof(dxxobject) * (num_objs + 1));
 					if (cur_objs == NULL) {
-						Warning ("Couldn't get %ld bytes for objects in interpolate playback\n", sizeof(dxxobject) * num_objs);
+						Warning ("Couldn't get %ld bytes for objects in interpolate playback\n", (unsigned long)(sizeof(dxxobject) * num_objs));
 						break;
 					}
 					for (objnum_t i = object_first; i <= num_objs; i++)
